+++ yugioh-back/src/main/java/com/naicson/yugioh/repository/CardRepository.java
@@ -10,17 +10,18 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import com.naicson.yugioh.dto.cards.CardAndSetsDTO;
+import com.naicson.yugioh.dto.cards.CardOfUserDetailDTO;
 import com.naicson.yugioh.dto.cards.CardsSearchDTO;
 import com.naicson.yugioh.entity.Card;
 import com.naicson.yugioh.util.CardSpecification;
 
 @Repository
-public interface CardRepository extends JpaRepository<Card, Integer>, JpaSpecificationExecutor<Card> {
+public interface CardRepository extends JpaRepository<Card, Long>, JpaSpecificationExecutor<Card> {
 	
 	List<Card> findAll();
 	Card findById(int id);
 	Card save (Card card);
-	Card findByNumero(Integer numero);
+	Card findByNumero(Long numero);
 	
 	void delete (Card card);
 	
@@ -46,5 +47,6 @@ public interface CardRepository extends JpaRepository<Card, Integer>, JpaSpecifi
 			countQuery = "SELECT count(*) FROM yugioh.tab_cards",
 			nativeQuery=true)
 	Page<Card> getByGenericType (Pageable page, String genericType, Integer userId);
+	
 	
 }
